🏗️ Architect: Modularized Solar Searches
- Refactored `apts/skyfield_searches/solar.py` into a modular package `apts/skyfield_searches/solar/`.
- Split logic into `eclipses.py` (lunar and solar eclipses) and `seasons.py` (solstices and equinoxes).
- Decomposed the complex `find_solar_eclipses` function into smaller private helpers `_classify_solar_eclipse` and `_calculate_solar_obscuration` to improve readability and reduce cyclomatic complexity.
- Maintained backward compatibility via re-exports in `__init__.py`.
