+++ apts/skyfield_searches/solar/__init__.py
@@ -1,0 +1,4 @@
+from .eclipses import find_lunar_eclipses, find_solar_eclipses
+from .seasons import find_seasons
+
+__all__ = ["find_lunar_eclipses", "find_solar_eclipses", "find_seasons"]
